Refactor text.py
